requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
assemblyai==0.15.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
Brotli==1.0.9
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
curl-cffi==0.5.7
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
dm-tree==0.1.8
exceptiongroup==1.1.2
executing==1.2.0
fake-useragent==1.1.3
ffmpeg==1.4
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.4.0
g4f==0.0.1.3
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
keras==2.12.0
kiwisolver==1.4.4
Levenshtein==0.21.1
libclang==16.0.0
lz4==4.3.2
MailGw-Temporary-Email==0.0.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==9.1.0
multidict==6.0.4
mutagen==1.46.0
names==0.3.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.6.26
platformdirs==3.8.0
prompt-toolkit==3.0.38
protobuf==4.23.3
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.0.3
pydantic_core==2.3.0
pydeck==0.8.1b0
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.15.1
pymailtm==1.1.1
Pympler==1.0.1
pyparsing==3.1.0
PyPasser==0.0.5
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-Levenshtein==0.21.1
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
Quandl==3.7.0
random-password-generator==2.2.0
random-username==1.0.2
rapidfuzz==3.1.1
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.10.0
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.8.1
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.24.1
streamlit-chat @ https://github.com/AI-Yash/st-chat/archive/refs/pull/24/head.zip
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.20.1
termcolor==2.3.0
threadpoolctl==3.1.0
tls-client==0.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
TwoCaptcha==0.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@1fa8b86f0b95f2e1488042ceeda8f356ea2a5448
youtube-transcript-api==0.6.1
yt-dlp==2023.7.6
zipp==3.16.2
